Refactored the dispatching algorithm
Doesn't build nested lists; it might still repeat subscriptions,
depending on how a process subscribes to topics (i.e. both subscribing
to "a/#" and "a/b/c" would result in the process being registered
twice) but this is handled by dropping the result list into a
MapSet. This could perhaps be optimized by properly handling of
subscriptions, so a process would only be represented once for topic
subscriptions that overshadows.